Split CI into discrete jobs (#981)
* Split CI into discrete jobs

* Update outdated POT file

* Skip writing POT file if the only thing changing is the date

* Format yaml

* Add yaml to lint staged

* Get locales when setting up node env

* Cram it all into a single workflow

* Add getting translations

+++ src/locales/scripts/json-to-pot.js
@@ -2,8 +2,21 @@ const json = require('../en.json')
 const fs = require('fs')
 const { createPotFile } = require('./json-pot-helpers')
 
+const matchPotCreationDate = /^"POT-Creation-Date: .*\\n"$/gm
+
 try {
   const fileName = process.cwd() + '/src/locales/po-files/openverse.pot'
+  const existingPotFile = String(fs.readFileSync(fileName)).replace(
+    matchPotCreationDate,
+    ''
+  )
+  const potFile = createPotFile(json)
+
+  if (existingPotFile === potFile.replace(matchPotCreationDate, '')) {
+    console.log('No change detected in pot file, skipping write')
+    return process.exit(0)
+  }
+
   fs.writeFileSync(fileName, createPotFile(json))
   console.log(`Successfully wrote pot file to ${fileName}`)
 } catch (err) {
